Move loggedInUserInfo from conversation state to app state
