+++ samples/apps/copilot-chat-app/webapp/src/redux/features/app/AppState.ts
@@ -3,6 +3,13 @@ import { AlertType } from '../../../libs/models/AlertType';
 // Copyright (c) Microsoft. All rights reserved.
 export interface AppState {
     alerts?: Alerts;
+    loggedInUserInfo?: LoggedInUserInfo;
+}
+
+export interface LoggedInUserInfo {
+    id: string;
+    email: string;
+    fullName: string;
 }
 
 export type Alert = {
