+++ samples/apps/copilot-chat-app/webapp/src/redux/features/app/appSlice.ts
@@ -2,7 +2,7 @@
 
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { AlertType } from '../../../libs/models/AlertType';
-import { Alert, Alerts, AppState } from './AppState';
+import { Alert, Alerts, AppState, LoggedInUserInfo } from './AppState';
 
 const initialState: AppState = {
     alerts: {
@@ -12,6 +12,7 @@ const initialState: AppState = {
             type: AlertType.Info,
         },
     },
+    loggedInUserInfo: undefined,
 };
 
 export const appSlice = createSlice({
@@ -33,9 +34,17 @@ export const appSlice = createSlice({
         removeAlert: (state: AppState, action: PayloadAction<string>) => {
             if (state.alerts) delete state.alerts[action.payload];
         },
+        setLoggedInUserInfo: (state: AppState, action: PayloadAction<LoggedInUserInfo>) => {
+            state.loggedInUserInfo = action.payload;
+        },
     },
 });
 
-export const { addAlert, removeAlert, setAlerts } = appSlice.actions;
+export const {
+    addAlert,
+    removeAlert,
+    setAlerts,
+    setLoggedInUserInfo
+} = appSlice.actions;
 
 export default appSlice.reducer;
