+++ samples/apps/copilot-chat-app/webapp/src/redux/features/conversations/ConversationsState.ts
@@ -10,19 +10,11 @@ export type Conversations = {
 export interface ConversationsState {
     conversations: Conversations;
     selectedId: string;
-    loggedInUserInfo?: LoggedInUserInfo;
-}
-
-export interface LoggedInUserInfo {
-    id: string;
-    email: string;
-    fullName: string;
 }
 
 export const initialState: ConversationsState = {
     conversations: {},
     selectedId: '',
-    loggedInUserInfo: undefined,
 };
 
 export type UpdateConversationPayload = {
